hide selector playground and studio toolbar when studio beta enabled
